requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
black==23.7.0
    # via -r requirements.in
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via black
colorama==0.4.6
    # via click
cryptography==41.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.2.3
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-auto-prefetch
    #   django-browser-reload
    #   django-filter
    #   django-markup
    #   django-taggit
    #   django-tailwind
    #   djangorestframework
    #   drf-yasg
    #   heroicons
django-allauth==0.54.0
    # via -r requirements.in
django-auto-prefetch==1.7.0
    # via -r requirements.in
django-browser-reload==1.11.0
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-markup==1.7.2
    # via -r requirements.in
django-robots==5.0
    # via -r requirements.in
django-simple-history==3.3.0
    # via -r requirements.in
django-taggit==4.0.0
    # via -r requirements.in
django-tailwind==3.6.0
    # via -r requirements.in
django-widget-tweaks==1.4.12
    # via -r requirements.in
djangorestframework==3.14.0
    # via drf-yasg
drf-yasg==1.21.7
    # via -r requirements.in
heroicons[django]==2.4.0
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-yasg
mypy-extensions==1.0.0
    # via black
mysqlclient==2.2.0
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   black
    #   drf-yasg
pathspec==0.11.2
    # via black
platformdirs==3.10.0
    # via black
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via django-allauth
python-dotenv==1.0.0
    # via -r requirements.in
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via drf-yasg
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
rules==3.3
    # via -r requirements.in
sqlparse==0.4.4
    # via django
taggit-selectize==2.11.0
    # via -r requirements.in
tomli==2.0.1
    # via black
typing-extensions==4.7.1
    # via asgiref
tzdata==2023.3
    # via django
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.0.4
    # via requests
zipp==3.16.2
    # via importlib-metadata
